+++ CHANGELOG.md
@@ -1,5 +1,13 @@
+# v0.2.0
+- Cleaner UX
+- Add command examples
+- Run tunnel as service.
+- See tunnel list
+- See ingress routes of a tunnel
+- Delete tunnel
+
 # v0.1.0
 - Create tunnel(s)
-- Add ingress rule(s) to tunnel
-- Remove ingress rule(s) from tunnel
+- Add ingress rule to tunnel
+- Remove ingress rule from tunnel
 - Run tunnel in foreground